Rename `utils.py` to `resource_search.py`.
